uservice-scheduler/requirements.txt
grpcio-reflection==1.62.0
dapr-ext-grpc==1.12.0
onlexnet_dapr == 0.0.2
onlexnet_scheduler == 0.0.7

# required as we have classes generated by avro-gen3
avro-gen3==0.7.12
fastavro==1.8.3

# required for tests
pytest==8.1.1
# required to run async tests
pytest-asyncio==0.23.6
